Create separate file for filter function, add function export to be able to mock data
