+++ src/js/modules/fetchData.js
@@ -24,16 +24,6 @@ export async function fetchScreenings(qs) {
   return data.data;
 }
 
-export async function filterSingleScreenings(id) {
-  const today = new Date().toISOString();
-
-  const obj = await fetchScreenings(`?filters[movie]=${id}`);
-  const filteredScreenings = [];
-
-  obj.forEach((obj) => {
-    if (obj.attributes.start_time >= today) filteredScreenings.push(obj);
-  });
-
-  const jsonScreenings = JSON.stringify(filteredScreenings);
-  return jsonScreenings;
-}
+export default {
+  fetchScreenings,
+};
