+++ src/js/routes/api.js
@@ -1,9 +1,13 @@
 import express from "express";
-import { filterSingleScreenings } from "../modules/fetchData.js";
+import { filterSingleScreenings } from "../modules/filterScreenings.js";
+import fetchFunc from "../modules/fetchData.js";
 const apiRouter = express.Router();
 
 apiRouter.get("/movies/:movieId/screenings", async (req, res) => {
-  const screenings = await filterSingleScreenings(req.params.movieId);
+  const screenings = await filterSingleScreenings(
+    req.params.movieId,
+    fetchFunc
+  );
   res.send(screenings);
 });
